requirements.txt
streamlit
google-generativeai
chromadb
pypdf
python-dotenv
unstructured
python-docx
pandas
openpyxl
